+++ src/js/td-scenes/battle/animation/game-state/battle/attack/genesis-braver/miss.ts
@@ -1,0 +1,21 @@
+import { Miss } from "gbraver-burst-core";
+
+import { Animate } from "../../../../../../../animation/animate";
+import { delay } from "../../../../../../../animation/delay";
+import { GenesisBraverBattle } from "./genesis-braver-battle";
+
+/**
+ * ミス
+ * @param param パラメータ
+ * @returns アニメーション
+ */
+export function miss(param: GenesisBraverBattle<Miss>): Animate {
+  return param.attackerSprite
+    .charge()
+    .chain(delay(500))
+    .chain(param.attackerSprite.straightPunch())
+    .chain(param.defenderSprite.avoid())
+    .chain(delay(500))
+    .chain(param.attackerSprite.spToStand())
+    .chain(delay(500));
+}
